Internal demuxing of transport streams encapsulated in t2mi or stid format.

+++ drivers/media/dvb-frontends/stid135/fesat_commlla_str.h
@@ -363,7 +363,7 @@ struct fe_sat_signal_info {
 	enum fe_sat_iq_inversion	spectrum;	/* IQ specrum swap setting				*/
 	u8				matype; //matype of current frame
 	bool low_roll_off_detected;
-	u8 				isi;		/* Current value of ISI 				*/
+	int 				isi;		/* Current value of ISI  or -1 for non-multistream or -2 for not initialized*/
 	u8        pls_mode;
 	u32       pls_code;
 	fe_sat_isi_struct isi_list;
